% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototypes.R
\name{block_urn_rand}
\alias{block_urn_rand}
\title{Block urn randomisation}
\usage{
block_urn_rand(target_alloc, sample_size, blocksize)
}
\arguments{
\item{target_alloc}{The target allocation ratio}

\item{sample_size}{The number of allocations to generate}

\item{blocksize}{The size of each block}
}
\value{
A list
}
\description{
Block urn randomisation
}
\references{
Zhao, W., & Weng, Y. (2011). Block urn design—a new randomization algorithm for sequential trials with two or more treatments and balanced or unbalanced allocation. Contemporary clinical trials, 32(6), 953-961.
}
