% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototypes.R
\name{complete_rand}
\alias{complete_rand}
\title{Complete randomisation}
\usage{
complete_rand(target_alloc, sample_size)
}
\arguments{
\item{target_alloc}{The target allocation ratios}

\item{sample_size}{The number of allocations to generate}
}
\value{
A list
}
\description{
Standard simple randomisation
}
