% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototypes.R
\name{modified_wei_urn_rand}
\alias{modified_wei_urn_rand}
\title{Modified Wei's Urn Design}
\usage{
modified_wei_urn_rand(target_alloc, sample_size, alpha, beta)
}
\arguments{
\item{target_alloc}{The target allocation ratios}

\item{sample_size}{The number of allocations to generate}
}
\value{
A list
}
\description{
Modified Wei's Urn Design
}
