% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototypes.R
\name{permuted_block_rand}
\alias{permuted_block_rand}
\title{Permuted block randomisation}
\usage{
permuted_block_rand(target_alloc, sample_size, blocksize)
}
\arguments{
\item{target_alloc}{The target allocation ratio}

\item{sample_size}{The number of allocations to generate}

\item{blocksize}{The size of each block}
}
\value{
A list
}
\description{
Permuted block randomisation
}
