% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandModel.R
\name{RandModel}
\alias{RandModel}
\title{Generalised Randomisation Model object}
\value{
Returns [R6::R6Class()] object of class RandModel
}
\description{
A generalised randomisation model object for defining custom randomisation schemes.
}
\details{
A generic randomisation model from which specific models inherit.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the RandModel}

\item{\code{short_name}}{The shorthand name of the RandModel}

\item{\code{description}}{A description of the RandModel}

\item{\code{target}}{The target allocation ratio/probabilities}

\item{\code{rng_seed}}{The integer seed to use for generating allocations}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RandModel$new()}}
\item \href{#method-print}{\code{RandModel$print()}}
\item \href{#method-get_name}{\code{RandModel$get_name()}}
\item \href{#method-get_description}{\code{RandModel$get_description()}}
\item \href{#method-get_target}{\code{RandModel$get_target()}}
\item \href{#method-get_history}{\code{RandModel$get_history()}}
\item \href{#method-get_rng}{\code{RandModel$get_rng()}}
\item \href{#method-num_arms}{\code{RandModel$num_arms()}}
\item \href{#method-num_allocations}{\code{RandModel$num_allocations()}}
\item \href{#method-randomise}{\code{RandModel$randomise()}}
\item \href{#method-clone}{\code{RandModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$new(
  name = NULL,
  short_name = NULL,
  description = NULL,
  target = NULL,
  history = NULL,
  rng_seed = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name.}

\item{\code{short_name}}{Short name.}

\item{\code{description}}{Description.}

\item{\code{target}}{Target allocation ratio}

\item{\code{history}}{Randomisation model history}

\item{\code{rng_seed}}{The RNG initial seed (integer)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `RandModel` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a RandModel.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Any additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-get_name}{}}}
\subsection{Method \code{get_name()}}{
Return the `RandModels` name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$get_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_description"></a>}}
\if{latex}{\out{\hypertarget{method-get_description}{}}}
\subsection{Method \code{get_description()}}{
Return the `RandModels` description
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$get_description()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_target"></a>}}
\if{latex}{\out{\hypertarget{method-get_target}{}}}
\subsection{Method \code{get_target()}}{
Return the target allocation for the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$get_target()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_history"></a>}}
\if{latex}{\out{\hypertarget{method-get_history}{}}}
\subsection{Method \code{get_history()}}{
Return the history of the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$get_history()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_rng"></a>}}
\if{latex}{\out{\hypertarget{method-get_rng}{}}}
\subsection{Method \code{get_rng()}}{
Return the state of the RNG of the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$get_rng()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-num_arms"></a>}}
\if{latex}{\out{\hypertarget{method-num_arms}{}}}
\subsection{Method \code{num_arms()}}{
Return the number of arms
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$num_arms()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-num_allocations"></a>}}
\if{latex}{\out{\hypertarget{method-num_allocations}{}}}
\subsection{Method \code{num_allocations()}}{
Return the number of allocations to each arm.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$num_allocations()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A table giving the number of allocations to each arm.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-randomise"></a>}}
\if{latex}{\out{\hypertarget{method-randomise}{}}}
\subsection{Method \code{randomise()}}{
Generate randomisations from the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$randomise(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{The number of assignments to generate}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of arm allocations of length `n`.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
