% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteRand.R
\name{CompleteRand}
\alias{CompleteRand}
\title{Complete Randomisation Model}
\value{
Returns R6 object of class CompleteRand
}
\description{
A complete randomisation model class
}
\section{Super class}{
\code{\link[rctrandr:RandModel]{rctrandr::RandModel}} -> \code{CompleteRand}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CompleteRand$new()}}
\item \href{#method-clone}{\code{CompleteRand$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="arm_names">}\href{../../rctrandr/html/RandModel.html#method-arm_names}{\code{rctrandr::RandModel$arm_names()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="conditional_prob">}\href{../../rctrandr/html/RandModel.html#method-conditional_prob}{\code{rctrandr::RandModel$conditional_prob()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="get_description">}\href{../../rctrandr/html/RandModel.html#method-get_description}{\code{rctrandr::RandModel$get_description()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="get_history">}\href{../../rctrandr/html/RandModel.html#method-get_history}{\code{rctrandr::RandModel$get_history()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="get_name">}\href{../../rctrandr/html/RandModel.html#method-get_name}{\code{rctrandr::RandModel$get_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="get_parameter_descriptions">}\href{../../rctrandr/html/RandModel.html#method-get_parameter_descriptions}{\code{rctrandr::RandModel$get_parameter_descriptions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="get_parameter_input_descriptions">}\href{../../rctrandr/html/RandModel.html#method-get_parameter_input_descriptions}{\code{rctrandr::RandModel$get_parameter_input_descriptions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="get_rng">}\href{../../rctrandr/html/RandModel.html#method-get_rng}{\code{rctrandr::RandModel$get_rng()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="get_short_name">}\href{../../rctrandr/html/RandModel.html#method-get_short_name}{\code{rctrandr::RandModel$get_short_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="get_target_allocation">}\href{../../rctrandr/html/RandModel.html#method-get_target_allocation}{\code{rctrandr::RandModel$get_target_allocation()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="num_allocations">}\href{../../rctrandr/html/RandModel.html#method-num_allocations}{\code{rctrandr::RandModel$num_allocations()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="num_arms">}\href{../../rctrandr/html/RandModel.html#method-num_arms}{\code{rctrandr::RandModel$num_arms()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="print">}\href{../../rctrandr/html/RandModel.html#method-print}{\code{rctrandr::RandModel$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="random_allocation">}\href{../../rctrandr/html/RandModel.html#method-random_allocation}{\code{rctrandr::RandModel$random_allocation()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of `CompleteRand` class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompleteRand$new(target_allocation = NULL, history = NULL, rng_seed = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{history}}{The model history of allocations. If NULL assumes no history.}

\item{\code{rng_seed}}{The RNG initial seed in the form of `.Random.seed`.}

\item{\code{target}}{A (named) numeric vector giving the target allocation ratio.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `CompleteRand` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompleteRand$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
