% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassWeightedUrnRand.R
\name{MassWeightedUrnRand}
\alias{MassWeightedUrnRand}
\title{Mass-weighted Urn Randomisation Model}
\value{
Returns R6 object of class [MassWeightedUrnRand].
}
\description{
A mass-weighted urn randomisation model class
}
\section{Super class}{
\code{\link[rctrandr:RandModel]{rctrandr::RandModel}} -> \code{MassWeightedUrnRand}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{imbalance_tolerance}}{Imbalance control parameter}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MassWeightedUrnRand$new()}}
\item \href{#method-conditional_prob}{\code{MassWeightedUrnRand$conditional_prob()}}
\item \href{#method-get_parameter_descriptions}{\code{MassWeightedUrnRand$get_parameter_descriptions()}}
\item \href{#method-clone}{\code{MassWeightedUrnRand$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="arm_names">}\href{../../rctrandr/html/RandModel.html#method-arm_names}{\code{rctrandr::RandModel$arm_names()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="get_description">}\href{../../rctrandr/html/RandModel.html#method-get_description}{\code{rctrandr::RandModel$get_description()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="get_history">}\href{../../rctrandr/html/RandModel.html#method-get_history}{\code{rctrandr::RandModel$get_history()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="get_name">}\href{../../rctrandr/html/RandModel.html#method-get_name}{\code{rctrandr::RandModel$get_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="get_parameter_input_descriptions">}\href{../../rctrandr/html/RandModel.html#method-get_parameter_input_descriptions}{\code{rctrandr::RandModel$get_parameter_input_descriptions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="get_rng">}\href{../../rctrandr/html/RandModel.html#method-get_rng}{\code{rctrandr::RandModel$get_rng()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="get_short_name">}\href{../../rctrandr/html/RandModel.html#method-get_short_name}{\code{rctrandr::RandModel$get_short_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="get_target_allocation">}\href{../../rctrandr/html/RandModel.html#method-get_target_allocation}{\code{rctrandr::RandModel$get_target_allocation()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="num_allocations">}\href{../../rctrandr/html/RandModel.html#method-num_allocations}{\code{rctrandr::RandModel$num_allocations()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="num_arms">}\href{../../rctrandr/html/RandModel.html#method-num_arms}{\code{rctrandr::RandModel$num_arms()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="print">}\href{../../rctrandr/html/RandModel.html#method-print}{\code{rctrandr::RandModel$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rctrandr" data-topic="RandModel" data-id="random_allocation">}\href{../../rctrandr/html/RandModel.html#method-random_allocation}{\code{rctrandr::RandModel$random_allocation()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of `MassWeightedUrnRand` class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassWeightedUrnRand$new(
  target_allocation = NULL,
  history = NULL,
  rng_seed = NULL,
  imbalance_tolerance = 5
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_allocation}}{The target allocation ratio}

\item{\code{history}}{The model history}

\item{\code{rng_seed}}{The RNG seed for the model}

\item{\code{imbalance_tolerance}}{The imbalance parameter}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `MassWeightedUrnRand` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-conditional_prob"></a>}}
\if{latex}{\out{\hypertarget{method-conditional_prob}{}}}
\subsection{Method \code{conditional_prob()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassWeightedUrnRand$conditional_prob()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_parameter_descriptions"></a>}}
\if{latex}{\out{\hypertarget{method-get_parameter_descriptions}{}}}
\subsection{Method \code{get_parameter_descriptions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassWeightedUrnRand$get_parameter_descriptions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassWeightedUrnRand$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
