% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandModel.R
\name{RandModel}
\alias{RandModel}
\title{Generalised Randomisation Model object}
\value{
Returns [R6::R6Class()] object of class RandModel
}
\description{
A generalised randomisation model object for defining custom randomisation schemes.
}
\details{
A generic randomisation model from which specific models inherit.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the [RandModel]}

\item{\code{short_name}}{The shorthand name of the [RandModel]}

\item{\code{description}}{A description of the [RandModel]}

\item{\code{target_allocation}}{The target allocation ratio/probabilities}

\item{\code{rng_seed}}{The integer seed to use for generating allocations}

\item{\code{history}}{The history of model allocations.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RandModel$new()}}
\item \href{#method-print}{\code{RandModel$print()}}
\item \href{#method-get_name}{\code{RandModel$get_name()}}
\item \href{#method-get_short_name}{\code{RandModel$get_short_name()}}
\item \href{#method-get_description}{\code{RandModel$get_description()}}
\item \href{#method-get_parameter_descriptions}{\code{RandModel$get_parameter_descriptions()}}
\item \href{#method-get_parameter_input_descriptions}{\code{RandModel$get_parameter_input_descriptions()}}
\item \href{#method-get_target_allocation}{\code{RandModel$get_target_allocation()}}
\item \href{#method-get_history}{\code{RandModel$get_history()}}
\item \href{#method-get_rng}{\code{RandModel$get_rng()}}
\item \href{#method-num_arms}{\code{RandModel$num_arms()}}
\item \href{#method-arm_names}{\code{RandModel$arm_names()}}
\item \href{#method-num_allocations}{\code{RandModel$num_allocations()}}
\item \href{#method-conditional_prob}{\code{RandModel$conditional_prob()}}
\item \href{#method-random_allocation}{\code{RandModel$random_allocation()}}
\item \href{#method-clone}{\code{RandModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$new(
  name = NULL,
  short_name = NULL,
  description = NULL,
  target_allocation = NULL,
  history = NULL,
  rng_seed = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name.}

\item{\code{short_name}}{Short name.}

\item{\code{description}}{Description.}

\item{\code{history}}{Randomisation model history}

\item{\code{rng_seed}}{The RNG initial seed (integer)}

\item{\code{target}}{Target allocation ratio}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `RandModel` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a RandModel.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Any additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-get_name}{}}}
\subsection{Method \code{get_name()}}{
The name of the randomisation model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$get_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string giving the model name.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_short_name"></a>}}
\if{latex}{\out{\hypertarget{method-get_short_name}{}}}
\subsection{Method \code{get_short_name()}}{
The short name of the randomisation model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$get_short_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string giving the model short name.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_description"></a>}}
\if{latex}{\out{\hypertarget{method-get_description}{}}}
\subsection{Method \code{get_description()}}{
The description of the randomisation model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$get_description()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string giving the model description.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_parameter_descriptions"></a>}}
\if{latex}{\out{\hypertarget{method-get_parameter_descriptions}{}}}
\subsection{Method \code{get_parameter_descriptions()}}{
The parameters required for the randomisation model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$get_parameter_descriptions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list giving the required model parameters.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_parameter_input_descriptions"></a>}}
\if{latex}{\out{\hypertarget{method-get_parameter_input_descriptions}{}}}
\subsection{Method \code{get_parameter_input_descriptions()}}{
The parameters required by the randomisation model to generate an allocation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$get_parameter_input_descriptions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list giving the required model inputs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_target_allocation"></a>}}
\if{latex}{\out{\hypertarget{method-get_target_allocation}{}}}
\subsection{Method \code{get_target_allocation()}}{
Return the target allocation for the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$get_target_allocation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_history"></a>}}
\if{latex}{\out{\hypertarget{method-get_history}{}}}
\subsection{Method \code{get_history()}}{
Return the history of the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$get_history()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_rng"></a>}}
\if{latex}{\out{\hypertarget{method-get_rng}{}}}
\subsection{Method \code{get_rng()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$get_rng()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The state of the RNG of the model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-num_arms"></a>}}
\if{latex}{\out{\hypertarget{method-num_arms}{}}}
\subsection{Method \code{num_arms()}}{
Return the number of arms.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$num_arms()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of arms in the model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-arm_names"></a>}}
\if{latex}{\out{\hypertarget{method-arm_names}{}}}
\subsection{Method \code{arm_names()}}{
Return the names of the arms
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$arm_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-num_allocations"></a>}}
\if{latex}{\out{\hypertarget{method-num_allocations}{}}}
\subsection{Method \code{num_allocations()}}{
Return the number of allocations to each arm.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$num_allocations()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A table giving the number of allocations to each arm.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-conditional_prob"></a>}}
\if{latex}{\out{\hypertarget{method-conditional_prob}{}}}
\subsection{Method \code{conditional_prob()}}{
Return the current conditional allocation probability given current state.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$conditional_prob()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A numeric vector with one element per arm giving that arms allocation probability.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random_allocation"></a>}}
\if{latex}{\out{\hypertarget{method-random_allocation}{}}}
\subsection{Method \code{random_allocation()}}{
Generate a single allocation from the randomisation model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$random_allocation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
